keyval: Fix Documentation, stash away pgfkeys experiments
The playground.tex and lilyglyph.tex documents are updated, 
so that they compile with the new version.

The main documentation has to be massively updated though to 
reflect the changes (NEXT TODO!!!)

I stashed away the experiments with pgfkeys.
I achieved the goal with keyval in the meantime, but it may still be
useful to change that to pdfkeys someday.