Fix #18 - Unwanted trailing space in \lilyText
When invoking \lilyPrint through \lilyText (i.e. with plain text as its argument) it introduced a trailing space.
This was fixed by removing the space in the effective printing command.

But as it turned out this caused an error when \lilyPrint was called with a single glyph (through \lilyGlyph).
So the solution was to introduce this space in the lowlevel commands \lilyGetGlyph and \lilyGetGlyphByNumber.