Introduce category subdirectories for image glyphs
As the directories containing the lily sources and pdfs for the image glyphs
would speedily get cluttered, I introduced category subdirectories.

The scripts will put the files in (possibly new) subdirectories of
/glyphimages/generated_src and /glyphimages/pdfs.
The name of the subdirectory corresponds to the input file, 
so an input file /glyphimages/definitions/singlenotes.ly will result in
'singlenotes' subdirectories.