Manual: Remove (comment out) glyphlist images:
I intended to incorporate the glyph list images from Lily's documentation
in the manual (in order to be able to see those glyphs that are possible 
but not yet implemented.
They take up very much space, result in awkward floating of the figures and tables,
and they make less sense now that we have the image commands.

For now the existing \glyphlist figures are commented out, but could eventually
be removed completely.