Manual: Reorder Chapters
Move Command Reference after Usage and before Internals.
Although the reference will grow longer and longer it seems like a good
idea to pull it more to the beginning of the manual.
For the average user it shouldn't be necessary to go through all the
internal stuff in order to reach the reference.