manual: Enable broken tables in \reftable
The \reftable used to document predefined commands is now
a {longtable}. This allows tables to be broken around pages
and to be placed directly where they are inserted in the text.

Floating ins't acceptable in this case because there are so many tables
that having them float through the text makes it _very_ impractical
to use the reference.