buildglyphimages: Make more versatile wrt directories
The script now expects a (relative of absolute) complete input file name.
This has to be in the definitions subdir of either the package dir or a copy of
the lilyglyphs_private structure.

The script now outputs the generated sources and the resulting pdfs in a single directory each,
not a hierarchy of folders corresponding to input file name.
This is because CTAN wants a hierarchy as flat as possible.

The generated latex file is put in a sibling folder (generated_cmd)
instead of a top-level 'stash' dir.