Scripts: import library also in a TeX installation
In a TeX installation the library file lilyglyphs_common.py won't be installed
in the same directory as the main script files.

This commit allows the scripts to either import the library directly or
use TeX's tool kpsewhich to locate the installed library path.

TODO: If the package is installed from the CTAN download in the personal texmf tree
kpsewhich won't find the library.
There has to be a solution for that too