Make font selection lua/xetex aware
Emmentaler fonts have to be in the texmf tree
(will be there when installed in a distro).

Xetex can find opentype fonts only by file name,
Luatex only by font name (arrrgh).

This commit lets both texes find it.

BUT: When a font is installed system-wide xetex only finds it
by font name (arrrrgh arrrrgh).

Therefore we have to tell the user to install the font in texmf.